restructured utility
