+++ src/backend/pyproject.toml
@@ -22,8 +22,6 @@ test = [
     "pytest>=7.3.1",
     "pytest-asyncio>=0.21.0"
 ]
-[tool.pytest.ini_options]
-pythonpath = ["."]
 
 [build-system]
 requires = ["setuptools>=61.0"]
