--- src/backend/requirements.txt
@@ -1,13 +1,0 @@
-# This file was autogenerated by uv via the following command:
-#    uv pip compile --project src/backend src/backend/pyproject.toml --no-deps
-azure-ai-inference==1.0.0b9
-azure-identity==1.21.0
-azure-monitor-opentelemetry-exporter==1.0.0b35
-fastapi==0.115.11
-openai==1.68.0
-opentelemetry-exporter-otlp-proto-grpc==1.31.1
-opentelemetry-instrumentation==0.52b1
-python-dotenv==1.0.1
-pyyaml==6.0.2
-semantic-kernel==1.25.0
-uvicorn==0.32.1
